requirements.txt
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
Django==2.2.4
django-mdeditor==0.1.18
djangocodemirror==2.1.0
djangorestframework==3.13.1
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
pycodestyle==2.8.0
Pygments==2.11.1
pytz==2021.3
some-package==0.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
